refactor(config)!: remove RMS locations from the default config
As this location can change with each release from AMX I have decided to
remove this from the default in app config.

The locations can still be added manually to either the global or a local
app config. However, it is recommended to symlink the required RMS files to the relevant AMXShare
locations.

For example symlink:
*.{axs,jar} to "C:\Program Files (x86)\Common Files\AMXShare\Duet\module"
*.axi to "C:\Program Files (x86)\Common Files\AMXShare\AXIs"